Move randomized integration test into its own file
